pyproject.toml
[tool.poetry]
name = "datain"
version = "0.1.0"
description = ""
authors = ["Mariano Ferro <ferro.mariano@gmail.com>"]
readme = "README.md"
packages = [{ include = "**/*" }]

[tool.poetry.dependencies]
python = "^3.10"
ccxt = "^4.3.98"
websockets = "^13.0.1"
websocket = "^0.2.1"
binance-futures-connector = "^4.0.0"
uvicorn = "^0.30.6"
fastapi = "^0.114.0"
rich = "^13.8.0"
poetry-dotenv = "^0.4.0"
pyjwt = "^2.9.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.10"
pytest = "^8.3.3"
load-dotenv = "^0.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
